// app/lib/types.ts
export interface Page {
// ...
  ctaSection?: {
    enabled: boolean;
    title?: any;
    description?: any;
    primaryButton?: {
      label: string;
      href: string;
    };
    backgroundImage?: string;
    backgroundColor?: string;
  };
// ...
}